Character Creation Data Handling
In this iteration, character creation was redefined. Firstly, the character class will be set to "Depraved" by default in invalid class names are chosen. Next I integrated a namedtuple in a calculate stats function that returns a different tuples based on the class selected. Thirdly, saving the character data to a save file is now possible. The option to load save files is also available. Error handling for read and write operations to save files is still ongoing.